pyproject.toml
[project]
name = "unusual_prompt"
version = "0.0.0"
description = "A Guardrails AI validator that validates a prompt for unusualness and trickery."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "guardrails-ai>=0.3.2",
    "litellm>=1.23.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]